+++ src/main/kotlin/com/hwalaon/wezxro_server/domain/account/model/CustomRate.kt
@@ -2,9 +2,9 @@ package com.hwalaon.wezxro_server.domain.account.model
 
 import com.hwalaon.wezxro_server.global.common.basic.model.BasicTime
 
-class CustomRate (
+class CustomRate(
     var id: Long?,
     var serviceId: Long?,
     var user: Account?,
-    var rate: Float
+    var rate: Double
 ): BasicTime()
